mv, config, utils, deepobj class to optics folder
