+++ 2_autolens_rms.py
@@ -18,7 +18,7 @@ from tqdm import tqdm
 
 from deeplens.geolens import GeoLens
 from deeplens.optics.geolens_pkg.utils import create_lens
-from deeplens.basics import DEPTH, EPSILON, WAVE_RGB
+from deeplens.optics.config import DEPTH, EPSILON, WAVE_RGB
 from deeplens.utils import create_video_from_images, set_logger, set_seed
