+++ deeplens/optics/geolens_pkg/eval.py
@@ -50,7 +50,7 @@ import numpy as np
 import torch
 import torch.nn.functional as F
 
-from deeplens.basics import (
+from deeplens.optics.config import (
     DEFAULT_WAVE,
     DEPTH,
     EPSILON,
