+++ deeplens/optics/geolens_pkg/tolerance.py
@@ -26,7 +26,7 @@ import numpy as np
 import torch
 from tqdm import tqdm
 
-from deeplens.basics import DEPTH
+from deeplens.optics.config import DEPTH
 
 
 class GeoLensTolerance:
