+++ deeplens/optics/geolens_pkg/view_3d.py
@@ -18,7 +18,7 @@ from typing import List, Optional
 import numpy as np
 import torch
 
-from deeplens.basics import DEFAULT_WAVE
+from deeplens.optics.config import DEFAULT_WAVE
 from deeplens.optics import Ray
 from deeplens.optics.geometric_surface import Aperture
