+++ deeplens/optics/phase_surface/nurbs.py
@@ -2,7 +2,7 @@
 
 import torch
 
-from deeplens.basics import EPSILON
+from deeplens.optics.config import EPSILON
 from deeplens.optics.phase_surface.phase import Phase
