+++ deeplens/optics/phase_surface/phase.py
@@ -5,7 +5,8 @@ import numpy as np
 import torch
 import torch.nn.functional as F
 
-from deeplens.basics import EPSILON, DeepObj
+from deeplens.optics.config import EPSILON
+from deeplens.optics import DeepObj
 from deeplens.optics.material import Material
