+++ deeplens/optics/phase_surface/poly.py
@@ -3,7 +3,7 @@
 import numpy as np
 import torch
 
-from deeplens.basics import EPSILON
+from deeplens.optics.config import EPSILON
 from deeplens.optics.phase_surface.phase import Phase
