+++ deeplens/paraxiallens.py
@@ -15,7 +15,7 @@ import numpy as np
 import torch
 
 from deeplens.lens import Lens
-from deeplens.basics import DEPTH, EPSILON, PSF_KS
+from deeplens.optics.config import DEPTH, EPSILON, PSF_KS
 from deeplens.optics.imgsim import conv_psf_depth_interp, conv_psf_occlusion
