+++ deeplens/psfnetlens.py
@@ -21,7 +21,7 @@ from deeplens.geolens import GeoLens
 from deeplens.lens import Lens
 from deeplens.network.surrogate import MLP
 from deeplens.network.surrogate.psfnet_mplconv import PSFNet_MLPConv
-from deeplens.basics import DEPTH, PSF_KS
+from deeplens.optics.config import DEPTH, PSF_KS
 from deeplens.optics.imgsim import conv_psf_pixel, conv_psf_pixel_high_res, rotate_psf
