+++ test/test_ray.py
@@ -6,7 +6,7 @@ import pytest
 import torch
 
 from deeplens.optics.light import Ray
-from deeplens.basics import DEFAULT_WAVE
+from deeplens.optics.config import DEFAULT_WAVE
 
 
 class TestRayInit:
